Rope is generating properly now.
Rope can be generated by inputting the desired number of links, link length, and link width. Next steps are to make a moveable body that can drag the end of the rope around. After that make the moveable end be able to extend the rope or retract the rope upon button press.